Modularize the structure to support Edition of events
