+++ app/src/main/java/com/github/se/gatherspot/model/EventViewModel.kt
@@ -1,14 +1,20 @@
 package com.github.se.gatherspot.model
 
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.github.se.gatherspot.EventFirebaseConnection
+import com.github.se.gatherspot.EventFirebaseConnection.Companion.fetchEvent
 import com.github.se.gatherspot.model.event.Event
 import com.github.se.gatherspot.model.event.EventStatus
 import com.github.se.gatherspot.model.location.Location
+import kotlinx.coroutines.async
+import kotlinx.coroutines.launch
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
 
-class EventViewModel {
+class EventViewModel : ViewModel() {
+
 
   /**
    * Create an event from verified data
@@ -193,6 +199,31 @@ class EventViewModel {
         parsedMinAttendees,
         parsedDateLimitInscription,
         parsedTimeLimitInscription)
+  }
+
+    /**
+     * Fetch an event from the database and return an event object
+     *
+     *
+     * @param eventID: The unique identifier of the event
+     * @return The event fetched
+     */
+    suspend fun fetchEvent(eventID: String): Event {
+        var fetchedEvent : Event? = null
+        viewModelScope.launch {
+            fetchedEvent = EventFirebaseConnection.fetchEvent(eventID)
+        }.join()
+
+        return fetchedEvent!!
+    }
+
+
+    /**
+     * Once an event is edited, fetch the event from the database, update the modified fields,
+     * get the unmodified data and update the event in the database
+     */
+    fun editEvent(eventID: String, newEvent: Event){
+
   }
 
   fun validateDate(date: String, eMessage: String): LocalDate {
